Refactor BinaryCompactObjectHelpers.
Moves a metafunction out of BinaryCompactObject.
