+++ src/Domain/Creators/BinaryCompactObjectHelpers.hpp
@@ -1,0 +1,23 @@
+// Distributed under the MIT License.
+// See LICENSE.txt for details.
+
+#pragma once
+
+#include <type_traits>
+
+namespace domain::creators::detail {
+// If `Metavariables` has a `domain` member struct and
+// `domain::enable_time_dependent_maps` is `true`, then
+// inherit from `std::true_type`; otherwise, inherit from `std::false_type`.
+template <typename Metavariables, typename = std::void_t<>>
+struct enable_time_dependent_maps : std::false_type {};
+
+template <typename Metavariables>
+struct enable_time_dependent_maps<Metavariables,
+                                  std::void_t<typename Metavariables::domain>>
+    : std::bool_constant<Metavariables::domain::enable_time_dependent_maps> {};
+
+template <typename Metavariables>
+constexpr bool enable_time_dependent_maps_v =
+    enable_time_dependent_maps<Metavariables>::value;
+}  // namespace domain::creators::detail
